test: move tests to __tests__/unit
