+++ __tests__/unit/categories/js/commitlint/commitlint.config.spec.ts
