+++ __tests__/unit/categories/js/commitlint/commitlint.entrypoint.spec.ts
