+++ __tests__/unit/categories/js/dependabot/dependabot.config.spec.ts
