+++ __tests__/unit/categories/js/dependabot/dependabot.entrypoint.spec.ts
