+++ __tests__/unit/categories/js/docker/docker.config.spec.ts
