+++ __tests__/unit/categories/js/docker/docker.entrypoint.spec.ts
