+++ __tests__/unit/categories/js/eslint/eslint.config.spec.ts
