+++ __tests__/unit/categories/js/husky/husky.entrypoint.spec.ts
