+++ __tests__/unit/categories/js/husky/husky.utils.spec.ts
