+++ __tests__/unit/categories/js/jest/jest.entrypoint.spec.ts
