+++ __tests__/unit/categories/js/lint-staged/lint-staged.config.spec.ts
