+++ __tests__/unit/categories/js/lint-staged/lint-staged.entrypoint.spec.ts
