+++ __tests__/unit/categories/js/lint-staged/lint-staged.utils.spec.ts
