+++ __tests__/unit/categories/js/prettier/prettier.config.spec.ts
