+++ __tests__/unit/categories/js/stylelint/stylelint.config.spec.ts
