+++ __tests__/unit/categories/js/stylelint/stylelint.entrypoint.spec.ts
