+++ __tests__/unit/categories/js/stylelint/stylelint.utils.spec.ts
