+++ __tests__/unit/states/categories.spec.ts
