+++ __tests__/unit/utils/installed.spec.ts
