+++ __tests__/unit/utils/npm.spec.ts
