+++ __tests__/unit/utils/prompt.spec.ts
